First steps towards building docker files for postres.  Initial tests look ok.

+++ standalone-metastore/src/main/sql/postgres/hive-schema-3.0.0.postgres.sql
@@ -631,7 +631,7 @@ CREATE TABLE "WM_POOL" (
     "RP_ID" bigint NOT NULL,
     "PATH" character varying(1024) NOT NULL,
     "PARENT_POOL_ID" bigint,
-    "ALLOC_FRACTION" DOUBLE,
+    "ALLOC_FRACTION" double precision,
     "QUERY_PARALLELISM" integer
 );
